+++ src/dacs/ofxLaserDacHelios.cpp
@@ -77,7 +77,7 @@ bool DacHelios:: sendFrame(const vector<Point>& points){
 		
 		if(lock()) {
 			frameMode = true;
-			for(int i = 0; i<points.size(); i++) {
+			for(size_t i= 0; i<points.size(); i++) {
 				
 				const Point& p2 = points[i];
 				p1.x = ofMap(p2.x,0,800, HELIOS_MIN, HELIOS_MAX);
@@ -137,7 +137,7 @@ bool DacHelios::sendPoints(const vector<Point>& points) {
 	if(lock()) {
 		frameMode = false;
 
-		for(int i = 0; i<points.size(); i++) {
+		for(size_t i= 0; i<points.size(); i++) {
 			const Point& p2 = points[i];
 			p1.x = ofMap(p2.x,0,800, HELIOS_MIN, HELIOS_MAX);
 			p1.y = ofMap(p2.y,800,0, HELIOS_MIN, HELIOS_MAX); // Y is UP
@@ -183,7 +183,7 @@ void DacHelios :: threadedFunction(){
 		// if we're out of points, let's replay the frame!
 		if(bufferedPoints.size()<samples_per_packet) {
 			if(frameMode && replayFrames) {
-				for(int i = 0; i<framePoints.size(); i++) {
+				for(size_t i= 0; i<framePoints.size(); i++) {
 					addPoint(framePoints[i]);
 				}
 				isReplaying = true;
