+++ src/shapes/ofxLaserManualShape.h
@@ -34,14 +34,14 @@ class ManualShape : public Shape{
 	}
 	void appendPointsToVector(vector<ofxLaser::Point>& destpoints, const RenderProfile& profile, float speedMultiplier) {
 		
-		for(int i = 0; i<points.size(); i++) {
+		for(size_t i = 0; i<points.size(); i++) {
 			destpoints.push_back(ofxLaser::Point(points[i], colours[i],1, useCalibration));
 		}
 	};
 	
 	
 	void addPreviewToMesh(ofMesh& mesh){
-		for(int i = 0; i<points.size();i++) {
+		for(size_t i = 0; i<points.size();i++) {
 			mesh.addVertex(points[i]);
 			mesh.addColor(colours[i]);
